fix(doc-sync): canonicalize native collaboration runtime
