+++ apps/doc-sync/src/__tests__/native-runtime.store.mjs
@@ -1,0 +1,16 @@
+export const rows = new Map()
+export let seed
+export function setSeed(page) { seed = page }
+export async function query(sql, params) {
+  if (sql.includes('SELECT ydoc FROM documents')) return { rows: rows.has(params[0]) ? [rows.get(params[0])] : [] }
+  if (sql.includes('SELECT page, name FROM saved_views')) return { rows: [{ page: seed, name: 'Native runtime' }] }
+  if (sql.includes('INSERT INTO documents')) {
+    rows.set(params[0], { ydoc: Buffer.from(params[1]), stateVector: Buffer.from(params[2]),
+      page: JSON.parse(params[3]), title: params[4], seq: (rows.get(params[0])?.seq ?? 0) + 1 })
+    return { rows: [] }
+  }
+  if (sql.includes('UPDATE saved_views')) return { rows: [] }
+  throw new Error(`Unexpected fixture SQL: ${sql}`)
+}
+export function queryWithRLS() { throw new Error('No user database in this fixture') }
+export function getPool() { throw new Error('No pool in this fixture') }
